requirements.txt
selenium
webdriver_manager
html-testRunner
pyyaml